Make the test setup more flexible.
The current tests fail with bucket size 2. We will need to modify the
tests so that they work for every (reasonable) bucket size before we can
change this value.

This change replaces the fixed close/mid/far contacts-in-buckets scheme
with one that allows for 256 contacts per bucket, so it will be easier
to adapt the tests later.