Only deal with XorNames and keep buckets separate.
This is a rewrite of most of the algorithms to keep a list of buckets
instead of a list of entries, which allows for considerable speedups in
some cases. The tests now run in 0.8 s instead of 5.7 s. (Admittedly,
some tests have been removed, but the slower ones are still there.)

It also removes the two generic arguments and makes RoutingTable only
keep names of contacts instead of NodeInfos.