feat/lib: make bucket size a parameter and name generic
This is a rewrite of most of the crate, to allow specifying a
`bucket_size` and `max_bucket_size`, independent of the group size used
for consensus.

For easier testing and greater flexibility, the routing table can use
types other than `XorName` as node names.

Some tests have been rewritten using `u8` or `u64` for greater speed and
readability.